+++ pkg/devicemapper/devmapper_wrapper.go
@@ -5,16 +5,7 @@ package devicemapper
 /*
 #cgo LDFLAGS: -L. -ldevmapper
 #include <libdevmapper.h>
-#include <linux/loop.h> // FIXME: present only for defines, maybe we can remove it?
 #include <linux/fs.h>   // FIXME: present only for BLKGETSIZE64, maybe we can remove it?
-
-#ifndef LOOP_CTL_GET_FREE
-  #define LOOP_CTL_GET_FREE 0x4C82
-#endif
-
-#ifndef LO_FLAGS_PARTSCAN
-  #define LO_FLAGS_PARTSCAN 8
-#endif
 
 // FIXME: Can't we find a way to do the logging in pure Go?
 extern void DevmapperLogCallback(int level, char *file, int line, int dm_errno_or_class, char *str);
@@ -45,44 +36,12 @@ import (
 
 type (
 	cdmTask C.struct_dm_task
-
-	loopInfo64 struct {
-		loDevice         uint64 /* ioctl r/o */
-		loInode          uint64 /* ioctl r/o */
-		loRdevice        uint64 /* ioctl r/o */
-		loOffset         uint64
-		loSizelimit      uint64 /* bytes, 0 == max available */
-		loNumber         uint32 /* ioctl r/o */
-		loEncryptType    uint32
-		loEncryptKeySize uint32 /* ioctl w/o */
-		loFlags          uint32 /* ioctl r/o */
-		loFileName       [LoNameSize]uint8
-		loCryptName      [LoNameSize]uint8
-		loEncryptKey     [LoKeySize]uint8 /* ioctl w/o */
-		loInit           [2]uint64
-	}
 )
 
 // IOCTL consts
 const (
 	BlkGetSize64 = C.BLKGETSIZE64
 	BlkDiscard   = C.BLKDISCARD
-
-	LoopSetFd       = C.LOOP_SET_FD
-	LoopCtlGetFree  = C.LOOP_CTL_GET_FREE
-	LoopGetStatus64 = C.LOOP_GET_STATUS64
-	LoopSetStatus64 = C.LOOP_SET_STATUS64
-	LoopClrFd       = C.LOOP_CLR_FD
-	LoopSetCapacity = C.LOOP_SET_CAPACITY
-)
-
-// LOOP consts.
-const (
-	LoFlagsAutoClear = C.LO_FLAGS_AUTOCLEAR
-	LoFlagsReadOnly  = C.LO_FLAGS_READ_ONLY
-	LoFlagsPartScan  = C.LO_FLAGS_PARTSCAN
-	LoKeySize        = C.LO_KEY_SIZE
-	LoNameSize       = C.LO_NAME_SIZE
 )
 
 // Devicemapper cookie flags.
